+++ app/stores/ada/HWVerifyAddressStore.js
@@ -1,6 +1,5 @@
 // @flow
 import { observable, action, toJS } from 'mobx';
-import { defineMessages } from 'react-intl';
 import Store from '../base/Store';
 
 import {
@@ -12,8 +11,7 @@ import {
   LedgerBridge,
 } from 'yoroi-extension-ledger-bridge';
 
-import LocalizableError, { UnexpectedError } from '../../i18n/LocalizableError';
-import globalMessages from '../../i18n/global-messages';
+import LocalizableError from '../../i18n/LocalizableError';
 import environment from '../../environment';
 
 import type {
@@ -23,17 +21,13 @@ import type {
 import {
   Logger,
   stringifyError,
-  // stringifyData,
 } from '../../utils/logging';
 
-import Wallet from '../../domain/Wallet';
+import {
+  convertToLocalizableError
+} from '../../domain/LedgerLocalizedError';
 
-const messages = defineMessages({
-  signTxError101: {
-    id: 'wallet.send.ledger.error.101',
-    defaultMessage: '!!!Signing cancelled on Ledger device. Please retry or reconnect device.',
-  },
-});
+import Wallet from '../../domain/Wallet';
 
 export default class AddressesStore extends Store {
   @observable isActionProcessing: boolean = false;
@@ -67,6 +61,7 @@ export default class AddressesStore extends Store {
 
     try {
       // TODO: don't use hardcoded strings and maybe find a better way to do this?
+      // $FlowFixMe
       if (params.wallet.hardwareInfo.vendor === 'ledger.com') {
         await this.ledgerVerifyAddress(path);
       } else if (params.wallet.hardwareInfo.vendor === 'trezor.io') {
@@ -77,7 +72,7 @@ export default class AddressesStore extends Store {
       }
     } catch (error) {
       Logger.error('AddressStore::_verifyAddress::error: ' + stringifyError(error));
-      this._setError(this._convertToLocalizableError(error));
+      this._setError(convertToLocalizableError(error));
     } finally {
       this._setActionProcessing(false);
     }
@@ -114,48 +109,10 @@ export default class AddressesStore extends Store {
       }
     } catch (error) {
       Logger.error('AddressStore::ledgerVerifyAddress::error: ' + stringifyError(error));
-      this._setError(this._convertToLocalizableError(error));
+      this._setError(convertToLocalizableError(error));
     } finally {
       Logger.info('HWVerifyStore::ledgerVerifyAddress finalized ');
     }
-  }
-
-  /**
-   * TODO: This needs to be moved to a single file (also from LedgerSendStore)
-   */
-
-  /** Converts error(from API or Ledger API) to LocalizableError */
-  _convertToLocalizableError = (error: any): LocalizableError => {
-    let localizableError: ?LocalizableError = null;
-
-    if (error instanceof LocalizableError) {
-      // It means some API Error has been thrown
-      localizableError = error;
-    } else if (error && error.message) {
-      // Ledger device related error happend, convert then to LocalizableError
-      switch (error.message) {
-        case 'TransportError: Failed to sign with Ledger device: U2F TIMEOUT':
-          localizableError = new LocalizableError(globalMessages.ledgerError101);
-          break;
-        case 'TransportStatusError: Ledger device: Action rejected by user':
-          localizableError = new LocalizableError(messages.signTxError101);
-          break;
-        default:
-          /** we are not able to figure out why Error is thrown
-            * make it, Something unexpected happened */
-          Logger.error(`LedgerSendStore::_convertToLocalizableError::error: ${error.message}`);
-          localizableError = new UnexpectedError();
-          break;
-      }
-    }
-
-    if (!localizableError) {
-      /** we are not able to figure out why Error is thrown
-        * make it, Something unexpected happened */
-      localizableError = new UnexpectedError();
-    }
-
-    return localizableError;
   }
 
   @action _selectAddress = async (params: { address: string, path: BIP32Path }): Promise<void> => {
